Improve search behavior and add &search command
Taxa are now dynamically searched and cached, removing the need for a hardcoded dictionary of supported taxa.

Queries are now passed to taxon_name instead of q, which prevents searching in other fields (such as location).

Observations now must have a minimum rank of species, preventing situations where an entire Class can be returned; this removes the need for a whitelist.

The search command "&search" has been added.